--- src/errors/index.ts
@@ -1,1 +1,0 @@
-export * from './app-error';
