+++ src/modules/auth/auth.service.ts
@@ -4,7 +4,7 @@ import * as crypto from 'crypto';
 import { StatusCodes } from 'http-status-codes';
 
 import { ErrorCode } from '~/enums';
-import { AppError } from '~/errors';
+import { HttpException } from '~/exceptions';
 import type { JwtService } from '~/modules/jwt';
 import { Role } from '~/modules/role';
 import type { UserService } from '~/modules/user';
@@ -31,7 +31,7 @@ export class AuthService {
       !user ||
       (user && !(onlyAdmin ? [Role.Admin] : [Role.User]).includes(user.roleId))
     ) {
-      throw new AppError(StatusCodes.NOT_FOUND, [
+      throw new HttpException(StatusCodes.NOT_FOUND, [
         {
           key: 'user',
           message: `User Not Found`,
@@ -41,7 +41,7 @@ export class AuthService {
     }
 
     if (user.provider !== AuthProviders.Email) {
-      throw new AppError(StatusCodes.BAD_REQUEST, [
+      throw new HttpException(StatusCodes.BAD_REQUEST, [
         {
           key: 'sign-in provider',
           message: `Need sign-in via ${user.provider} provider`,
@@ -56,7 +56,7 @@ export class AuthService {
     );
 
     if (!isValidPassword) {
-      throw new AppError(StatusCodes.BAD_REQUEST, [
+      throw new HttpException(StatusCodes.BAD_REQUEST, [
         {
           key: 'password',
           message: `Incorrect Password`,
